rename and move refactorimngs for proposal stuff
