+++ plugins/org.yakindu.sct.refactoring/META-INF/MANIFEST.MF
@@ -13,8 +13,7 @@ Require-Bundle: org.eclipse.ui,
  org.yakindu.sct.model.stext;bundle-version="1.0.0",
  org.yakindu.base.xtext.utils.jface;bundle-version="1.0.0",
  org.yakindu.sct.model.resource;bundle-version="1.0.0",
- org.eclipse.core.expressions,
- org.yakindu.base.xtext.utils.gmf;bundle-version="2.1.2"
+ org.eclipse.core.expressions
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-ActivationPolicy: lazy
 Export-Package: org.yakindu.sct.refactoring.proposals,
