+++ plugins/org.yakindu.sct.refactoring/src/org/yakindu/sct/refactoring/proposals/RefactoringBasedModification.java
@@ -13,9 +13,9 @@ package org.yakindu.sct.refactoring.proposals;
 import java.util.Collections;
 
 import org.eclipse.gmf.runtime.notation.View;
-import org.yakindu.base.xtext.utils.gmf.proposals.ISemanticModification;
 import org.yakindu.sct.refactoring.refactor.AbstractRefactoring;
 import org.yakindu.sct.refactoring.refactor.IRefactoring;
+import org.yakindu.sct.ui.editor.modifications.ISemanticModification;
 
 /**
  * Adapts the {@link ISemanticModification} interface to the
