+++ plugins/org.yakindu.sct.ui.editor/META-INF/MANIFEST.MF
@@ -60,6 +60,7 @@ Export-Package: org.yakindu.sct.ui.editor,
  org.yakindu.sct.ui.editor.editor.proposals,
  org.yakindu.sct.ui.editor.editparts,
  org.yakindu.sct.ui.editor.factories,
+ org.yakindu.sct.ui.editor.modifications,
  org.yakindu.sct.ui.editor.partitioning,
  org.yakindu.sct.ui.editor.policies,
  org.yakindu.sct.ui.editor.preferences,
