+++ plugins/org.yakindu.sct.ui.editor/src/org/yakindu/sct/ui/editor/editor/proposals/AddCompositeModification.java
@@ -11,6 +11,7 @@
 package org.yakindu.sct.ui.editor.editor.proposals;
 
 import org.eclipse.gmf.runtime.notation.View;
+import org.yakindu.sct.ui.editor.modifications.StateTemplatesModification;
 
 /**
  *
